apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Daily Dish FastAPI backend"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "authlib>=1.7.0",
    "cloudinary>=1.44.2",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.26",
    "sqlmodel>=0.0.38",
    "uuid6>=2025.0.1",
    "uvicorn>=0.46.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "greenlet>=3.4.0",
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.11",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
